+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.11] - Unreleased
+
+### Added
+
+- Added Table.show_lines to render lines between rows
+
+### Changed
+
+- Added markup escape with double square brackets
+
 ## [0.8.10] - 2020-04-12
 
 ### Fixed